3月6日までの作業量(深野)

+++ src/Assets/FukanoTest/BulletData.cs
@@ -19,11 +19,11 @@ public class BulletData
     public BulletType type;
     public string name;
     public int ID;
-    public float power;
+    public int power;
     public float recoil;
     public float angle;
     public float range;
-    public float Levelpower; 
+    public int Levelpower; 
     public float Levelrecoil;
     public float Levelangle;
     public float Levelrange;
